Dynamic API requests for unit pages
